Add files via upload
This program is created for the user to determine its Body Mass Index and Weight Height Ratio. It has a special features wherein the user can see some information regarding Body Mass Index (BMI) and Weight Height Ratio (WHR) like e.g what to do when the user is extremely obese and what nutrient patterns the user must follow in order to be in good shape (normal BMI and WHR). It also have a keypress function wherein the user can't input letters when the required information is in numbers.

The program is the created in C# program and it was done in Visual Studio Environment